+++ src/com/ansorgit/plugins/bash/editor/codefolding/BashVariableFoldingBuilder.kt
@@ -4,46 +4,48 @@ import com.ansorgit.plugins.bash.lang.lexer.BashTokenTypes
 import com.ansorgit.plugins.bash.lang.psi.api.vars.BashVar
 import com.ansorgit.plugins.bash.lang.psi.api.vars.BashVarDef
 import com.ansorgit.plugins.bash.lang.psi.impl.vars.BashVarDefImpl
+import com.ansorgit.plugins.bash.lang.psi.impl.vars.ConfigurableSmartVarReference
 import com.ansorgit.plugins.bash.lang.psi.stubs.elements.BashVarElementType
 import com.intellij.lang.ASTNode
 import com.intellij.lang.folding.FoldingBuilderEx
 import com.intellij.lang.folding.FoldingDescriptor
 import com.intellij.openapi.editor.Document
+import com.intellij.openapi.project.DumbAware
 import com.intellij.psi.PsiElement
 import com.intellij.psi.search.PsiElementProcessor
 import com.intellij.psi.util.PsiTreeUtil
 import com.intellij.util.containers.ContainerUtil.newArrayList
 
 /**
- * Folding for variables in bash script. 
- * 
- * For example this code: 
+ * Folding for variables in bash script.
+ *
+ * For example this code:
  *     PWD=/tmp/daily-nsnam/foo
  *     say "-d forcing PWD = $PWD"
- *     
+ *
  * getting like:
  *     PWD=/tmp/daily-nsnam/foo
  *     say "-d forcing PWD = /tmp/daily-nsnam/foo"
- * 
+ *
  * @author Nosov Aleksandr <nosovae.dev@gmail.com>
  */
-class BashVariableFoldingBuilder : FoldingBuilderEx() {
+class BashVariableFoldingBuilder : FoldingBuilderEx(), DumbAware {
 
     private val DEFAULT_DEPTH_OF_FOLDING = 5
-    private val DEFAULT_PLACEHOLDER = ""
 
     override fun getPlaceholderText(node: ASTNode) = getPlaceholderText(node, DEFAULT_DEPTH_OF_FOLDING)
 
     private fun getPlaceholderText(node: ASTNode, depth: Int): String {
         val bashVar = node.psi as BashVar
-        val reference = bashVar.reference.resolve()
+
+        val reference = ConfigurableSmartVarReference(bashVar.reference, true).resolve()
         if (reference != null && reference is BashVarDefImpl) {
             val value = reference.findAssignmentValue()
             if (value != null) {
                 return computePlaceholderText(value.node, depth)
             }
         }
-        return DEFAULT_PLACEHOLDER
+        return node.text
     }
 
     private fun computePlaceholderText(valueNode: ASTNode, currentDepth: Int): String {
@@ -71,7 +73,9 @@ class BashVariableFoldingBuilder : FoldingBuilderEx() {
         ))
 
         foldingBlocks.forEach { psiElement ->
-            descriptors.add(FoldingDescriptor(psiElement, psiElement.textRange))
+            if (psiElement.reference?.resolve() != null) {
+                descriptors.add(FoldingDescriptor(psiElement, psiElement.textRange))
+            }
         }
 
         return descriptors.toTypedArray()